[SandboxIR][NFC] Move sandboxir::Use into a separate file
This helps avoid circular dependencies in a follow-up patch.
